Types for Paginator component
